Reorganized files into groups
