+++ Source/Core/HCCore.h
