+++ Source/Core/HCObject.h
