+++ Source/Core/HCObject_Internal.h
