+++ Source/Data/HCNumber_Internal.h
@@ -9,7 +9,7 @@
 #ifndef HCNumber_Internal_h
 #define HCNumber_Internal_h
 
-#include "HCObject_Internal.h"
+#include "../Core/HCObject_Internal.h"
 #include "HCNumber.h"
 
 //----------------------------------------------------------------------------------------------------------------------------------
