+++ Source/Data/HCString.c
