+++ Source/HollowCore.h
@@ -9,13 +9,13 @@
 #ifndef HollowCore_h
 #define HollowCore_h
 
-#include "HCCore.h"
-#include "HCObject.h"
-#include "HCNumber.h"
-#include "HCString.h"
-#include "HCData.h"
-#include "HCList.h"
-#include "HCSet.h"
-#include "HCMap.h"
+#include "Core/HCCore.h"
+#include "Core/HCObject.h"
+#include "Data/HCNumber.h"
+#include "Data/HCString.h"
+#include "Data/HCData.h"
+#include "Container/HCList.h"
+#include "Container/HCSet.h"
+#include "Container/HCMap.h"
 
 #endif /* HollowCore_h */
